realse 2.0

+++ src/main/java/com/elearning/demo/category/CategoryController.java
@@ -1,5 +1,11 @@
 package com.elearning.demo.category;
 
+import com.elearning.demo.attempt.Attempt;
+import com.elearning.demo.question.Question;
+import com.elearning.demo.question.QuestionService;
+import com.elearning.demo.quiz.Quiz;
+import com.elearning.demo.quiz.QuizService;
+import com.elearning.demo.study.Study;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -10,6 +16,12 @@ import java.util.List;
 public class CategoryController {
     @Autowired
     private CategoryService categoryService;
+
+    @Autowired
+    private QuizService quizService;
+
+    @Autowired
+    private QuestionService questionService;
 
     @GetMapping("/admin/categories")
     public List<Category> categories() {return categoryService.findAllCategory();}
@@ -29,7 +41,25 @@ public class CategoryController {
     }
 
     @DeleteMapping("/admin/categories/{id}")
-    public void deleteStudy(@PathVariable(value = "id") Long id) {
+    public Category deleteStudy(@PathVariable(value = "id") Long id) {
+        Category category = categoryService.findCategoryById(id);
+        List<Question> categoryQuestions = category.getQuestions();
+//        for (Question question: categoryQuestions) {
+//            question.setCategory(null);
+//            questionService.saveQuestion(question);
+//        }
+        List<Quiz> categoryQuizzes = category.getQuizzes();
+//        for (Quiz quiz: categoryQuizzes) {
+//            quiz.setCategory(null);
+//            quizService.saveQuiz(quiz);
+//        }
+        if (categoryQuestions.size() != 0) {
+            return null;
+        }
+        if (categoryQuizzes.size() != 0) {
+            return null;
+        }
         categoryService.removeCategory(id);
+        return category;
     }
 }
